requirements.txt
ortools>=9.10.0
PyYAML>=6.0
Flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=20.0.0